requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.3.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.3.2
cron-descriptor==1.4.5
Django==4.2.7
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
factory-boy==3.3.0
Faker==37.4.2
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
openpyxl==3.1.2
boto3==1.35.0
django-storages==1.14.2
